second commit

+++ setup.py
@@ -3,7 +3,7 @@ import os
 
 class Setup():
     def __init__(self):
-        self.log_path = 'botlog.log'
+        self.log_path = './output/botlog.log'
 
 
     def setup_log(self):
@@ -25,6 +25,7 @@ class Setup():
 
 
     def clean_screenshots_dir(self):
-        for file in os.listdir('./screenshots'):
-            file_path = os.path.join('./screenshots', file)
-            os.remove(file_path)
+        for file in os.listdir('./output'):
+            file_path = os.path.join('./output', file)
+            if file.lower().endswith('.png'):
+                os.remove(file_path)
